requirements.txt
streamlit
ollama
streamlit-pdf-viewer
PyPDF2
pdfplumber
PyMuPDF
pytest
pytest-cov
pillow 